elf2flt: clean up elf headers dependencies
The only reason we included cygwin-elf.h & elf.h was for the various
target reloc defines.  However, since these are all provided by the
bfd library which we already include, we can switch to using that for
every target.

Now we don't have any dep on the host's ELF headers (either existing,
or being up-to-date).