Revert "elf2flt: fix for segfault on some ARM ELFs"
This reverts commit ba379d08bb78c9300e84351c11080c26ddcc36b3.

This change breaks at least the m68k architecture support. It causes
overlapping text and data sections resulting in errors like this:

    ERROR: text=0x38f34 overlaps data=0x31a74 ?

The underlying cause here is that the .eh_frame section is marked
as readonly relocated data on m68k, resulting in it being treated
now as if it will be loaded into the flat text section. But it
is layed out in the linker script as part of the flat data section.

Note that this is actually a partial revert. It preserves the debug
printf fix in the original commit.

Signed-of-by: Greg Ungerer <gerg@kernel.org>