Revert "elf2flt: fix relocations for read-only data"
This reverts commit 73325b7f209e0f68887333385184af275531427d.

Although this fixed the simple problem case (elf2flt core dumping on
binutils newer than 2.33.1) it causes further breakage that required
further hacks to fix that resulted in more breakage.

Back all of these changes out and find a better way.

Note that this is actually a partial revert. It preserves the comment
fix in the original commit.

Signed-off-by: Greg Ungerer <gerg@kernel.org>