elf2flt: force ARM.exidx section into text
Fix problem with el2flt core dumping on binutils version 2.33.1 and
newer. Previous attempts at fixing this have had a list of subtle
problems causing further breakage.

The problem stems from the .ARM.exidx section being marked as load
readonly data, but the elf2flt.ld linker script puts it into the
flat text section. The elf2flt processing code expects all separate
ELF sections marked as "data" to be in the flat data section and
calculates sizing and memory based on that. (Note that traditional
readonly data is treated as text in the linker script, so it is not
affected here and is merged into the flat text section with no
problems).

The retaining of separate ELF sections outside of the flat text and
flat data flatmem sections proper is fragile. Fundamentally the
flat format only has the two stored sections - text and data. By far
the best appraoch is to explicilty put everything into those named
sections in the elf2flt.ld.

The .ARM.exidx section must be part of the flat text section - due
to the relocations that it carries. So move it into the text section
proper in the elf2flt.ld linker script.

Signed-off-by: Greg Ungerer <gerg@kernel.org>