elf2flt: remove use of BFD_VMA_FMT
In binutils-2.40 the BFD_VMA_FMT definition used for printf style
formatting specifiers has been removed. For reference this was done in
commit b82817674f46 ("Don't use BFD_VMA_FMT in binutils") in the
binutils git development tree.

BFD_VMA_FMT is used in a number of places in the elf2flt code to output
bfd offsets, values and the like. So it is broken when using the bfd
code from binutils-2.40 and newer.

According to the binutils change PRIx64 (and friends) is used to replace
it, with appropriate casts to keep it clean for 32 and 64 bit platforms.
Change the elf2flt.c use of it in the same way to fix.

This does not change the output in any way in normal use. This fix can
be used on all versions of binutils (older and newer), there is no
need to only do this on 2.40 and newer.

Signed-off-by: Greg Ungerer <gerg@kernel.org>